better organize code and add jumptables test(currently passes)
