+++ src/app/setup.h
@@ -1,0 +1,6 @@
+#ifndef DOLRECOMP_APP_SETUP_H
+#define DOLRECOMP_APP_SETUP_H
+
+int run_setup(void);
+
+#endif
