+++ src/common/types.h
