+++ tests/test_rel.c
@@ -1,5 +1,5 @@
-#include "frontend/rel.h"
-#include "core/types.h"
+#include "frontend/container/rel.h"
+#include "common/types.h"
 #include <stdio.h>
 #include <string.h>
