+++ tests/test_rpx.c
@@ -1,5 +1,5 @@
-#include "frontend/rpx.h"
-#include "core/types.h"
+#include "frontend/container/rpx.h"
+#include "common/types.h"
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
